requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-functions==1.13.3
azure-functions-worker==1.1.9
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
grpcio==1.33.2
grpcio-tools==1.33.2
idna==3.4
multidict==6.0.4
openai==0.27.4
protobuf==3.20.3
requests==2.28.2
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.9.1
